Committing as-is WIP, it compiles...
Want to get a backup of the work I've done so far as I pause for a bit.
This code compiles but has not been tested once.  I'd recommend not
using it in this state.

Added in some hashicorp vault code for handling secrets stored in
vault.  Lots of this is experimental, not even sure if this is the
final approach I wish to take.

Updated launch.json to remove a bunch of cli arguments that should no
longer be needed

Updated main.go with the additional vault code that I've added in.